requirements.txt
numpy==2.2.6
openai==1.82.1
python-dotenv==1.1.0
pyttsx3==2.98
pyaudio